examples/pyproject.toml
[tool.poetry]
name = "examples"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7,<3.12"
momento-python-redis-client = { path = "../../momento-python-redis-client", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
